docs: document aspect helpers
