+++ docs/.vitepress/config.ts
@@ -46,6 +46,7 @@ export default withMermaid(
               { text: 'Magic Proxy System', link: '/guide/magic-proxy' },
               { text: 'JavaScript to CEL', link: '/guide/javascript-to-cel' },
               { text: 'Deployment Modes', link: '/guide/deployment-modes' },
+              { text: 'Aspects', link: '/guide/aspects' },
               { text: 'Resource Scopes & Lifecycle', link: '/guide/resource-scopes' },
               { text: 'External References', link: '/guide/external-references' },
               { text: 'Type Safety Patterns', link: '/guide/type-safety' },
@@ -60,6 +61,7 @@ export default withMermaid(
               { text: 'Overview', link: '/api/' },
               { text: 'Import Patterns', link: '/api/imports' },
               { text: 'kubernetesComposition', link: '/api/kubernetes-composition' },
+              { text: 'Aspects', link: '/api/aspects' },
               { text: 'CEL Expressions', link: '/api/cel' },
               { text: 'YAML & Helm Integration', link: '/api/yaml-closures' },
               { text: 'Container Build', link: '/api/containers/' },
