+++ src/core/aspects/index.ts
@@ -18,609 +18,29 @@
  * fields; `append(...)` is valid for advertised array fields. Kro mode rejects
  * `merge(...)` and `append(...)` when either the current composite field or the
  * operation payload contains Kubernetes refs or CEL expressions.
- *
- * @example Apply metadata across a rendered composition.
- * ```typescript
- * app.toYaml({
- *   aspects: [
- *     aspect.on(allResources, metadata({ labels: merge({ team: 'platform' }) })),
- *   ],
- * });
- * ```
- *
- * @example Override factory-advertised workload fields for hot reload.
- * ```typescript
- * app.toYaml({
- *   aspects: [
- *     aspect
- *       .on(simple.Deployment, override({
- *         spec: {
- *           template: {
- *             spec: {
- *               containers: append([{
- *                 name: 'dev-tools',
- *                 image: 'busybox',
- *                 command: ['sh', '-c', 'sleep infinity'],
- *                 workingDir: '/workspace',
- *                 env: [{ name: 'LOG_LEVEL', value: 'debug' }],
- *                 volumeMounts: [{ name: 'workspace', mountPath: '/workspace' }],
- *               }],
- *               volumes: append([{ name: 'workspace', emptyDir: {} }]),
- *             },
- *           },
- *         },
- *       }))
- *       .where({ slot: 'web', labels: { tier: 'app' } })
- *       .expectOne(),
- *   ],
- * });
- * ```
- *
- * @example Apply the same aspect through direct/Kro factories.
- * ```typescript
- * const aspects = [aspect.on(allResources, metadata({ annotations: merge({ owner: 'platform' }) }))];
- * app.factory('direct', { aspects }).toYaml(spec);
- * app.factory('kro', { aspects }).toYaml();
- * ```
- *
- * Failure recovery: `AspectDefinitionError` means the aspect definition is invalid
- * before resources are touched (for example, wrong target/surface pairing).
- * `AspectApplicationError` means the definition matched resources but could not be
- * safely applied (for example, unsupported writable field, wrong operation type,
- * selector cardinality mismatch, or Kro reference-backed merge/append). Fix the
- * target, selector, field path, or operation and render again. Error diagnostics
- * include selector and resource identity context but not operation payloads or
- * full manifests.
  */
-import {
-  isCelExpression,
-  isKubernetesRef,
-  isMixedTemplate,
-  isResourceReference,
-} from '../../utils/type-guards.js';
-import { getAspectMetadata, setAspectMetadata } from './metadata.js';
-import { resolveFactoryTargetId } from './targets.js';
-import { AspectDefinitionError } from './types.js';
-
-import type {
-  V1EnvFromSource,
-  V1EnvVar,
-  V1ResourceRequirements,
-  V1Volume,
-  V1VolumeMount,
-} from '@kubernetes/client-node';
-
-import type {
-  AnnotationMap,
-  AppendOperation,
-  AspectBuilder,
-  AspectDefinition,
-  AspectDefinitionFunctionName,
-  AspectOverridePatch,
-  AspectSelector,
-  AspectSurface,
-  AspectSurfaceForTarget,
-  AspectTarget,
-  AspectTargetGroup,
-  CommonAspectSurfaceForTargets,
-  CompatibleAspectTargets,
-  EnvVarMap,
-  HotReloadAspectOptions,
-  HotReloadAspectSchema,
-  ImagePullPolicy,
-  LabelMap,
-  LocalWorkspaceAspectOptions,
-  MergeOperation,
-  MergeByNameOperation,
-  MetadataAspectSurface,
-  OverrideAspectSurface,
-  PatchEachOperation,
-  ReplaceOperation,
-  WorkloadPodTemplateAspectSchema,
-} from './types.js';
-
-function isPlainObject(value: unknown): value is Record<string, unknown> {
-  return typeof value === 'object' && value !== null && !Array.isArray(value);
-}
-
-function cloneImmutable<T>(value: T): T {
-  if (
-    isKubernetesRef(value) ||
-    isCelExpression(value) ||
-    isMixedTemplate(value) ||
-    isResourceReference(value)
-  ) {
-    return value;
-  }
-
-  if (Array.isArray(value)) {
-    return Object.freeze(value.map((entry) => cloneImmutable(entry))) as T;
-  }
-
-  if (isPlainObject(value)) {
-    const cloned: Record<string, unknown> = {};
-    for (const [key, child] of Object.entries(value)) {
-      cloned[key] = cloneImmutable(child);
-    }
-    return Object.freeze(cloned) as T;
-  }
-
-  return value;
-}
-
-function assertSelector(selector: AspectSelector): void {
-  if (!isPlainObject(selector)) {
-    throw createDefinitionError('aspect.on', 'where(...) selector must be an object');
-  }
-  const allowedKeys = new Set(['slot', 'id', 'name', 'namespace', 'kind', 'labels']);
-  for (const [key, value] of Object.entries(selector)) {
-    if (!allowedKeys.has(key)) {
-      throw createDefinitionError('aspect.on', `where(...) selector field ${key} is not supported`);
-    }
-    if (key === 'labels') {
-      if (!isPlainObject(value)) {
-        throw createDefinitionError('aspect.on', 'where(...) selector labels must be an object');
-      }
-      for (const [labelKey, labelValue] of Object.entries(value)) {
-        if (typeof labelValue !== 'string') {
-          throw createDefinitionError(
-            'aspect.on',
-            `where(...) selector label ${labelKey} must be a string`
-          );
-        }
-      }
-    } else if (typeof value !== 'string') {
-      throw createDefinitionError('aspect.on', `where(...) selector ${key} must be a string`);
-    }
-  }
-}
-
-function createDefinitionError(
-  functionName: AspectDefinitionFunctionName,
-  reason: string
-): AspectDefinitionError {
-  return new AspectDefinitionError(functionName, reason);
-}
-
-function operationKind(value: unknown): string | undefined {
-  return isPlainObject(value) && typeof value.kind === 'string' ? value.kind : undefined;
-}
-
-function assertMetadataOperation(field: string, value: unknown): void {
-  const kind = operationKind(value);
-  if (kind !== 'merge' && kind !== 'replace') {
-    throw createDefinitionError(
-      'metadata',
-      `metadata.${field} must use merge(...) or replace(...)`
-    );
-  }
-}
-
-function assertOverridePatch(path: string, value: unknown): void {
-  const kind = operationKind(value);
-  if (kind) {
-    if (kind === 'merge' && !isPlainObject((value as { value?: unknown }).value)) {
-      throw createDefinitionError('override', `${path} merge value must be an object`);
-    }
-    if (kind === 'append' && !Array.isArray((value as { value?: unknown }).value)) {
-      throw createDefinitionError('override', `${path} append value must be an array`);
-    }
-    if (kind === 'mergeByName' && !Array.isArray((value as { value?: unknown }).value)) {
-      throw createDefinitionError('override', `${path} mergeByName value must be an array`);
-    }
-    if (kind === 'patchEach') {
-      const patch = (value as { patch?: unknown }).patch;
-      if (!isPlainObject(patch)) {
-        throw createDefinitionError('override', `${path} patchEach patch must be an object`);
-      }
-      assertOverridePatch(`${path}.patch`, patch);
-      return;
-    }
-    if (kind === 'replace' || kind === 'merge' || kind === 'append' || kind === 'mergeByName') {
-      return;
-    }
-    throw createDefinitionError('override', `${path} uses unsupported operation ${kind}`);
-  }
-
-  if (!isPlainObject(value)) {
-    throw createDefinitionError('override', `${path} must contain aspect operations`);
-  }
-
-  for (const [key, child] of Object.entries(value)) {
-    assertOverridePatch(`${path}.${key}`, child);
-  }
-}
-
-function assertOverrideRoot(patch: Record<string, unknown>): void {
-  const keys = Object.keys(patch);
-  if (keys.length !== 1 || keys[0] !== 'spec') {
-    throw createDefinitionError('override', 'override(...) patch must be rooted at spec');
-  }
-  if (!isPlainObject(patch.spec)) {
-    throw createDefinitionError('override', 'override(...) spec patch must be an object');
-  }
-}
-
-function assertTargetSurfaceCompatibility(
-  targetOrTargets: AspectTarget | readonly AspectTarget[],
-  surface: AspectSurface
-): void {
-  const targets = Array.isArray(targetOrTargets) ? targetOrTargets : [targetOrTargets];
-  for (const target of targets) {
-    if (isPlainObject(target) && target.kind === 'target-group') {
-      if (target.id === 'allResources' && surface.kind !== 'metadata') {
-        throw createDefinitionError('aspect.on', 'allResources supports metadata(...) only');
-      }
-      if ((target.id === 'resources' || target.id === 'workloads') && surface.kind !== 'override') {
-        throw createDefinitionError(
-          'aspect.on',
-          `${target.id} supports override(...) aspects only`
-        );
-      }
-      continue;
-    }
-
-    if (typeof target === 'function') {
-      const targetId = resolveFactoryTargetId(target);
-      if (targetId === undefined) {
-        throw createDefinitionError(
-          'aspect.on',
-          'factory target must be a registered TypeKro factory or advertise TypeKro aspect metadata'
-        );
-      }
-      const surfaces = Reflect.get(target, '__typekroAspectSurfaces');
-      if (Array.isArray(surfaces) && !surfaces.includes(surface.kind)) {
-        throw createDefinitionError(
-          'aspect.on',
-          `factory target ${targetId} does not support ${surface.kind}(...) aspects`
-        );
-      }
-      continue;
-    }
-
-    throw createDefinitionError(
-      'aspect.on',
-      'target must be an aspect-capable factory or target group'
-    );
-  }
-}
-
-class AspectDefinitionImpl<TTarget = AspectTarget, TSurface = AspectSurface> {
-  readonly kind = 'aspect';
-
-  readonly selector?: AspectSelector;
-
-  constructor(
-    readonly target: TTarget,
-    readonly surface: TSurface,
-    readonly cardinality: AspectDefinition<TTarget, TSurface>['cardinality'] = 'one-or-more',
-    selector?: AspectSelector
-  ) {
-    if (selector !== undefined) {
-      this.selector = cloneImmutable(selector);
-    }
-    Object.freeze(this);
-  }
-
-  where(selector: AspectSelector): AspectDefinition<TTarget, TSurface> {
-    if (this.selector !== undefined) {
-      throw createDefinitionError('aspect.on', 'where(...) cannot be called more than once');
-    }
-    assertSelector(selector);
-    return new AspectDefinitionImpl(
-      this.target,
-      this.surface,
-      this.cardinality,
-      cloneImmutable(selector)
-    );
-  }
-
-  optional(): AspectDefinition<TTarget, TSurface> {
-    if (this.cardinality === 'exactly-one') {
-      throw createDefinitionError('aspect.on', 'optional() cannot be called after expectOne()');
-    }
-    return new AspectDefinitionImpl(this.target, this.surface, 'zero-or-more', this.selector);
-  }
-
-  expectOne(): AspectDefinition<TTarget, TSurface> {
-    if (this.cardinality === 'zero-or-more') {
-      throw createDefinitionError('aspect.on', 'expectOne() cannot be called after optional()');
-    }
-    return new AspectDefinitionImpl(this.target, this.surface, 'exactly-one', this.selector);
-  }
-}
-
-export const allResources: AspectTargetGroup<'allResources'> = Object.freeze({
-  kind: 'target-group',
-  id: 'allResources',
-});
-
-export const resources: AspectTargetGroup<'resources'> = Object.freeze({
-  kind: 'target-group',
-  id: 'resources',
-});
-
-export const workloads: AspectTargetGroup<'workloads'> = Object.freeze({
-  kind: 'target-group',
-  id: 'workloads',
-});
-
-function aspectOn<TTarget extends AspectTarget>(
-  target: TTarget,
-  surface: AspectSurfaceForTarget<TTarget>
-): AspectDefinition<TTarget, AspectSurfaceForTarget<TTarget>>;
-function aspectOn<const TTargets extends readonly AspectTarget[]>(
-  targets: CompatibleAspectTargets<TTargets>,
-  surface: CommonAspectSurfaceForTargets<TTargets>
-): AspectDefinition<TTargets, CommonAspectSurfaceForTargets<TTargets>>;
-function aspectOn(
-  targetOrTargets: AspectTarget | readonly AspectTarget[],
-  surface: AspectSurface
-): AspectDefinition {
-  assertTargetSurfaceCompatibility(targetOrTargets, surface);
-  return new AspectDefinitionImpl(targetOrTargets, surface) as unknown as AspectDefinition;
-}
-
-export const aspect: AspectBuilder = Object.freeze({
-  on: aspectOn,
-});
-
-/** Creates a replace operation descriptor. */
-export function replace<T>(value: T): ReplaceOperation<T> {
-  return Object.freeze({ kind: 'replace', value: cloneImmutable(value) });
-}
-
-/** Creates an object merge operation descriptor. */
-export function merge<T extends object>(value: Partial<T>): MergeOperation<T> {
-  if (!isPlainObject(value)) {
-    throw createDefinitionError('merge', 'merge(...) value must be a concrete object');
-  }
-  return Object.freeze({ kind: 'merge', value: cloneImmutable(value) });
-}
-
-/** Creates an array append operation descriptor. */
-export function append<TElement>(value: readonly TElement[]): AppendOperation<TElement> {
-  if (!Array.isArray(value)) {
-    throw createDefinitionError('append', 'append(...) value must be an array');
-  }
-  return Object.freeze({ kind: 'append', value: cloneImmutable(value) });
-}
-
-function patchEach<TElement extends object>(
-  patch: AspectOverridePatch<TElement>
-): PatchEachOperation<TElement> {
-  if (!isPlainObject(patch)) {
-    throw createDefinitionError('override', 'patchEach patch must be an object');
-  }
-  return Object.freeze({ kind: 'patchEach', patch: cloneImmutable(patch) });
-}
-
-function mergeByName<TElement extends { name: string }>(
-  value: readonly TElement[]
-): MergeByNameOperation<TElement> {
-  if (!Array.isArray(value)) {
-    throw createDefinitionError('override', 'mergeByName value must be an array');
-  }
-  for (const entry of value) {
-    if (!isPlainObject(entry) || typeof entry.name !== 'string' || entry.name.length === 0) {
-      throw createDefinitionError('override', 'mergeByName entries must have a non-empty name');
-    }
-  }
-  return Object.freeze({ kind: 'mergeByName', value: cloneImmutable(value) });
-}
-
-/** Creates a metadata aspect surface descriptor. */
-export function metadata(surface: Omit<MetadataAspectSurface, 'kind'>): MetadataAspectSurface {
-  if (!isPlainObject(surface)) {
-    throw createDefinitionError('metadata', 'metadata(...) value must be an object');
-  }
-  for (const key of Object.keys(surface)) {
-    if (key !== 'labels' && key !== 'annotations') {
-      throw createDefinitionError('metadata', `metadata.${key} is not supported`);
-    }
-  }
-  if (surface.labels !== undefined) assertMetadataOperation('labels', surface.labels);
-  if (surface.annotations !== undefined)
-    assertMetadataOperation('annotations', surface.annotations);
-  return Object.freeze({ kind: 'metadata', ...cloneImmutable(surface) });
-}
-
-/** Creates a resource override aspect surface descriptor. */
-export function override<TSchema extends object>(
-  patch: AspectOverridePatch<TSchema>
-): OverrideAspectSurface<TSchema>;
-export function override(patch: unknown) {
-  if (!isPlainObject(patch)) {
-    throw createDefinitionError('override', 'override(...) patch must be an object');
-  }
-  assertOverrideRoot(patch);
-  assertOverridePatch('patch', patch);
-  return Object.freeze({
-    kind: 'override',
-    patch: cloneImmutable(patch) as AspectOverridePatch<object>,
-  });
-}
-
-/** Creates a cross-cutting resource metadata label aspect. */
-export function withLabels(labels: LabelMap): AspectDefinition<typeof allResources, MetadataAspectSurface> {
-  return aspect.on(allResources, metadata({ labels: merge(labels) }));
-}
-
-/** Creates a cross-cutting resource metadata annotation aspect. */
-export function withAnnotations(
-  annotations: AnnotationMap
-): AspectDefinition<typeof allResources, MetadataAspectSurface> {
-  return aspect.on(allResources, metadata({ annotations: merge(annotations) }));
-}
-
-/** Creates a cross-cutting resource metadata aspect. */
-export function withMetadata(options: {
-  readonly labels?: LabelMap;
-  readonly annotations?: AnnotationMap;
-}): AspectDefinition<typeof allResources, MetadataAspectSurface> {
-  if (!isPlainObject(options)) {
-    throw createDefinitionError('metadata', 'withMetadata(...) options must be an object');
-  }
-  return aspect.on(
-    allResources,
-    metadata({
-      ...(options.labels !== undefined && { labels: merge(options.labels) }),
-      ...(options.annotations !== undefined && { annotations: merge(options.annotations) }),
-    })
-  );
-}
-
-function workloadOverride(
-  patch: AspectOverridePatch<WorkloadPodTemplateAspectSchema>
-): AspectDefinition<typeof workloads, OverrideAspectSurface<WorkloadPodTemplateAspectSchema>> {
-  return aspect.on(workloads, override<WorkloadPodTemplateAspectSchema>(patch)) as unknown as AspectDefinition<
-    typeof workloads,
-    OverrideAspectSurface<WorkloadPodTemplateAspectSchema>
-  >;
-}
-
-/** Creates a workload aspect that sets replica count. */
-export function withReplicas(
-  count: number
-): AspectDefinition<typeof workloads, OverrideAspectSurface<WorkloadPodTemplateAspectSchema>> {
-  return workloadOverride({ spec: { replicas: replace(count) } });
-}
-
-/** Creates a workload aspect that adds or updates env vars on every container. */
-export function withEnvVars(
-  vars: EnvVarMap
-): AspectDefinition<typeof workloads, OverrideAspectSurface<WorkloadPodTemplateAspectSchema>> {
-  const env: V1EnvVar[] = Object.entries(vars).map(([name, value]) => ({ name, value }));
-  return workloadOverride({
-    spec: { template: { spec: { containers: patchEach({ env: mergeByName(env) }) } } },
-  });
-}
-
-/** Creates a workload aspect that appends envFrom sources to every container. */
-export function withEnvFrom(
-  envFrom: readonly V1EnvFromSource[]
-): AspectDefinition<typeof workloads, OverrideAspectSurface<WorkloadPodTemplateAspectSchema>> {
-  return workloadOverride({
-    spec: { template: { spec: { containers: patchEach({ envFrom: append(envFrom) }) } } },
-  });
-}
-
-/** Creates a workload aspect that sets resource requirements on every container. */
-export function withResourceDefaults(
-  resources: V1ResourceRequirements
-): AspectDefinition<typeof workloads, OverrideAspectSurface<WorkloadPodTemplateAspectSchema>> {
-  return workloadOverride({
-    spec: { template: { spec: { containers: patchEach({ resources: replace(resources) }) } } },
-  });
-}
-
-/** Creates a workload aspect that sets imagePullPolicy on every container. */
-export function withImagePullPolicy(
-  policy: ImagePullPolicy
-): AspectDefinition<typeof workloads, OverrideAspectSurface<WorkloadPodTemplateAspectSchema>> {
-  return workloadOverride({
-    spec: { template: { spec: { containers: patchEach({ imagePullPolicy: replace(policy) }) } } },
-  });
-}
-
-/** Creates a workload aspect that sets serviceAccountName on pod templates. */
-export function withServiceAccount(
-  serviceAccountName: string
-): AspectDefinition<typeof workloads, OverrideAspectSurface<WorkloadPodTemplateAspectSchema>> {
-  return workloadOverride({
-    spec: { template: { spec: { serviceAccountName: replace(serviceAccountName) } } },
-  });
-}
-
-/** Creates a workload aspect that mounts a local host workspace into every container. */
-export function withLocalWorkspace(
-  options: LocalWorkspaceAspectOptions
-): AspectDefinition<typeof workloads, OverrideAspectSurface<WorkloadPodTemplateAspectSchema>> {
-  if (!isPlainObject(options)) {
-    throw createDefinitionError('override', 'withLocalWorkspace(...) options must be an object');
-  }
-  const volumeName = options.volumeName ?? 'workspace';
-  const mountPath = options.mountPath ?? '/workspace';
-  const volumeMount: V1VolumeMount & { name: string } = { name: volumeName, mountPath };
-  const volume: V1Volume & { name: string } = {
-    name: volumeName,
-    hostPath: { path: options.workspacePath, type: options.hostPathType ?? 'Directory' },
-  };
-
-  return workloadOverride({
-    spec: {
-      template: {
-        spec: {
-          containers: patchEach({ volumeMounts: mergeByName([volumeMount]) }),
-          volumes: mergeByName([volume]),
-        },
-      },
-    },
-  });
-}
-
-/**
- * Creates a dev-mode hot reload override surface for workload pod templates.
- *
- * The helper intentionally returns only the override surface. Callers still pick
- * the target and selector with `aspect.on(...)`, keeping dev behavior explicit.
- */
-export function hotReload(options: HotReloadAspectOptions): OverrideAspectSurface<HotReloadAspectSchema> {
-  if (!isPlainObject(options)) {
-    throw createDefinitionError('hotReload', 'hotReload(...) options must be an object');
-  }
-  if (!Array.isArray(options.containers) || options.containers.length === 0) {
-    throw createDefinitionError('hotReload', 'hotReload(...) containers must be a non-empty array');
-  }
-  if (options.volumes !== undefined && !Array.isArray(options.volumes)) {
-    throw createDefinitionError('hotReload', 'hotReload(...) volumes must be an array');
-  }
-  if (options.labels !== undefined && !isPlainObject(options.labels)) {
-    throw createDefinitionError('hotReload', 'hotReload(...) labels must be an object');
-  }
-
-  const labelPatch: Record<string, ReplaceOperation<string>> = {};
-  for (const [key, value] of Object.entries(options.labels ?? {})) {
-    labelPatch[key] = replace(value);
-  }
-
-  return override<HotReloadAspectSchema>({
-    spec: {
-      ...(options.replicas !== undefined && { replicas: replace(options.replicas) }),
-      template: {
-        ...(options.labels !== undefined && { metadata: { labels: labelPatch } }),
-        spec: {
-          containers: replace(options.containers),
-          ...(options.volumes !== undefined && { volumes: replace(options.volumes) }),
-        },
-      },
-    },
-  });
-}
-
-/** Creates a workload-targeted dev-mode hot reload aspect. */
-export function withHotReload(
-  options: HotReloadAspectOptions
-): AspectDefinition<typeof workloads, OverrideAspectSurface<HotReloadAspectSchema>> {
-  return aspect.on(workloads, hotReload(options)) as unknown as AspectDefinition<
-    typeof workloads,
-    OverrideAspectSurface<HotReloadAspectSchema>
-  >;
-}
-
-/** Attaches semantic slot metadata to a resource for aspect selector matching. */
-export function slot<TResource extends object>(name: string, resource: TResource): TResource {
-  if (name.length === 0) {
-    throw createDefinitionError('slot', 'slot name must not be empty');
-  }
-  const existingSlot = getAspectMetadata(resource)?.slot;
-  if (existingSlot !== undefined && existingSlot !== name) {
-    throw createDefinitionError('slot', 'resource already has a different slot assigned');
-  }
-  setAspectMetadata(resource, { slot: name });
-  return resource;
-}
-
 export { AspectApplicationError, applyAspects } from './apply.js';
+export { hotReload, withHotReload, withLocalWorkspace } from './dev-aspects.js';
+export { withAnnotations, withLabels, withMetadata } from './metadata-aspects.js';
+export {
+  allResources,
+  append,
+  aspect,
+  merge,
+  metadata,
+  override,
+  replace,
+  resources,
+  slot,
+  workloads,
+} from './primitives.js';
 export { AspectDefinitionError } from './types.js';
 export type * from './types.js';
+export {
+  withEnvFrom,
+  withEnvVars,
+  withImagePullPolicy,
+  withReplicas,
+  withResourceDefaults,
+  withServiceAccount,
+} from './workload-aspects.js';
